It now can convert walls strings to lists of booleans and back
It will convert the wall strings to bool list when generating the gameboard by default, can be changed with an argument in the generateBoard function.
Wall strings look like this: "tr" and they represent the directions where the cell has a wall ("t" = top, "r" = right etc.). When converting it to a list of bools, it creates something like this: [True, True, False, False], which tells you if there is a wall at each of the directions (the list represents the directions like this: [top, right, bottom, left])